# src/main/java/com/urlshortener/model/DeviceTypeStat.java
package com.urlshortener.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class DeviceTypeStat {
    private String deviceType;
    private long clicks;
}