# src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: URL Shortener Service API
  version: 1.0.0
  description: |
    REST API for generating and managing short URLs, with redirection and
    click-level analytics support.

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: URLs
    description: Short URL management
  - name: Analytics
    description: Click analytics and device distribution

paths:

  /v1/url/create:
    post:
      tags: [URLs]
      summary: Create a short URL
      operationId: createUrl
      description: |
        Creates a new short URL. Provide a custom alias (6–20 chars) or omit it
        for an auto-generated 8-character code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUrlRequest'
            examples:
              custom-alias:
                summary: Request with a custom alias
                value:
                  longUrl: "https://www.example.com/very/long/path?query=value"
                  alias: "my-link"
              auto-alias:
                summary: Request without an alias (auto-generated)
                value:
                  longUrl: "https://www.example.com/another/long/path"
      responses:
        '201':
          description: Short URL created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUrlResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Requested alias is already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/url:
    get:
      tags: [URLs]
      summary: List all short URLs (paginated)
      operationId: listUrls
      description: Returns a paginated list of all short URL records. Page size is fixed at 10.
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page number (default 0)
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list of short URLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortUrlPageResponse'
        '400':
          description: Invalid page parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /{shortCode}:
    get:
      tags: [URLs]
      summary: Redirect to the original long URL
      operationId: redirectToUrl
      description: |
        Resolves the short code to its original long URL and responds with an
        HTTP 302 redirect. A click event is recorded asynchronously.
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            example: "abc123"
      responses:
        '302':
          description: Redirect to the original long URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          description: Short code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/analytics:
    get:
      tags: [Analytics]
      summary: Get analytics for a 7-day window
      operationId: getAnalytics
      description: |
        Returns click analytics for a selected 7-day window.
        page=0 is the most recent; page=1 is the preceding one, etc.
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based window index (0 = last 7 days)
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Analytics for the selected window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          description: Invalid page parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/analytics/devices:
    get:
      tags: [Analytics]
      summary: Get device-type distribution
      operationId: getDeviceDistribution
      description: |
        Returns click counts grouped by device type
        (DESKTOP, MOBILE, TABLET, BOT, UNKNOWN) for the last 30 days.
      responses:
        '200':
          description: Device-type distribution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDistributionResponse'

components:
  schemas:

    # ── Requests ───────────────────────────────────────────────────────────────

    CreateUrlRequest:
      type: object
      required: [longUrl]
      properties:
        longUrl:
          type: string
          format: uri
          description: The original URL to shorten
          example: "https://www.example.com/very/long/path?query=value"
        alias:
          type: string
          minLength: 6
          maxLength: 20
          description: |
            Optional custom alias (6–20 characters). If omitted, an 8-character
            alphanumeric alias is generated automatically.
          example: "my-link"

    # ── URL responses ──────────────────────────────────────────────────────────

    CreateUrlResponse:
      type: object
      description: Response returned after successfully creating a short URL
      properties:
        id:
          type: integer
          format: int64
          description: Auto-generated primary key of the short URL record
          example: 42
        shortCode:
          type: string
          description: The alias/short code portion of the URL
          example: "abc123"
        shortUrl:
          type: string
          format: uri
          description: The fully-qualified short URL ready to share
          example: "http://localhost:8080/abc123"
        longUrl:
          type: string
          format: uri
          description: The original long URL
          example: "https://www.example.com/very/long/path"
        domain:
          type: string
          description: The domain used for this short URL
          example: "localhost:8080"
        aliasType:
          type: string
          enum: [CUSTOM, GENERATED]
          description: Whether the alias was user-supplied or auto-generated
          example: "GENERATED"
        createdAt:
          type: string
          format: date-time
          description: UTC ISO-8601 timestamp when the record was created
          example: "2025-03-08T12:00:00Z"

    ShortUrlSummary:
      type: object
      description: Summary of a single short URL entry
      properties:
        alias:
          type: string
          example: "aB3xK9mZ"
        longUrl:
          type: string
          format: uri
          example: "https://www.example.com/some/long/path"
        shortUrl:
          type: string
          format: uri
          example: "http://localhost:8080/aB3xK9mZ"
        clickCount:
          type: integer
          format: int64
          description: Total number of clicks recorded for this short URL
          example: 17

    ShortUrlPageResponse:
      type: object
      description: Paginated list of short URL summaries
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ShortUrlSummary'
        page:
          type: integer
          description: Zero-based current page number
          example: 0
        pageSize:
          type: integer
          description: Fixed page size
          example: 10
        totalRecords:
          type: integer
          format: int64
          description: Total number of short URL records
          example: 42
        totalPages:
          type: integer
          description: Total number of pages
          example: 5

    # ── Analytics responses ────────────────────────────────────────────────────

    AnalyticsSummary:
      type: object
      description: Click-count comparison between the current and previous 7-day windows
      properties:
        currentPeriodClicks:
          type: integer
          format: int64
          example: 128
        previousPeriodClicks:
          type: integer
          format: int64
          example: 95
        percentChange:
          type: number
          format: double
          description: Percentage change (positive = increase)
          example: 34.74

    AnalyticsPagination:
      type: object
      description: Pagination metadata for analytics windows
      properties:
        currentPage:
          type: integer
          description: Zero-based index of the selected 7-day window
          example: 0
        totalPages:
          type: integer
          description: Total number of available 7-day windows
          example: 4

    TimeseriesPoint:
      type: object
      description: Click count for a single calendar day
      properties:
        date:
          type: string
          format: date
          description: ISO-8601 date (UTC)
          example: "2025-03-08"
        clicks:
          type: integer
          format: int64
          example: 23

    AnalyticsResponse:
      type: object
      description: Full analytics response for a selected 7-day window
      properties:
        summary:
          $ref: '#/components/schemas/AnalyticsSummary'
        timeseries:
          type: array
          items:
            $ref: '#/components/schemas/TimeseriesPoint'
        pagination:
          $ref: '#/components/schemas/AnalyticsPagination'

    DeviceTypeStat:
      type: object
      description: Click count for a single device type
      properties:
        deviceType:
          type: string
          enum: [DESKTOP, MOBILE, TABLET, BOT, UNKNOWN]
          example: "MOBILE"
        clicks:
          type: integer
          format: int64
          example: 54

    DeviceDistributionResponse:
      type: object
      description: Device-type click distribution for the last 30 days
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DeviceTypeStat'

    # ── Error ──────────────────────────────────────────────────────────────────

    ErrorResponse:
      type: object
      description: Standard error response body
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Human-readable error description
          example: "Short URL 'xyz' not found."